etc: 파일명, 디렉토리 정리
